+++ frontend/src/components/calendarPage/index.ts
@@ -1,0 +1,11 @@
+export { CalendarPage } from './CalendarPage'
+export { CalendarHeader } from './CalendarHeader'
+export { CalendarGrid } from './CalendarGrid'
+export { CalendarDay } from './CalendarDay'
+export { SelectedDateMeals } from './SelectedDateMeals'
+export { RecentActivity } from './RecentActivity'
+
+// Hooks
+export { useCalendarData } from './hooks/useCalendarData'
+export { useMealOperations } from './hooks/useMealOperations'
+export { useMealPlanGeneration } from './hooks/useMealPlanGeneration'
